+++ src/models/Entity.php
@@ -40,6 +40,11 @@ class Entity extends FluentModel
     public $siteSpecificType;
 
     /**
+     * @var string The computed most specific schema.org type
+     */
+    public $computedType = 'Organization';
+
+    /**
      * @var string
      */
     public $genericName;
@@ -60,7 +65,22 @@ class Entity extends FluentModel
     public $genericUrl;
 
     /**
-     * @var array
+     * @var array URL for the entity image
+     */
+    public $genericImage;
+
+    /**
+     * @var int The width of the entity image
+     */
+    public $genericImageWidth;
+
+    /**
+     * @var int The height of the entity image
+     */
+    public $genericImageHeight;
+
+    /**
+     * @var array Asset ID array for the entity image
      */
     public $genericImageIds;
 
@@ -193,10 +213,12 @@ class Entity extends FluentModel
                     'siteType',
                     'siteSubType',
                     'siteSpecificType',
+                    'computedType',
                     'genericName',
                     'genericAlternateName',
                     'genericDescription',
                     'genericUrl',
+                    'genericImage',
                     'genericTelephone',
                     'genericEmail',
                     'genericStreetAddress',
@@ -211,6 +233,7 @@ class Entity extends FluentModel
                     'organizationDuns',
                     'organizationFounder',
                     'organizationFoundingDate',
+                    'organizationFoundingLocation',
                     'corporationTickerSymbol',
                     'localBusinessPriceRange',
                     'restaurantServesCuisine',
@@ -218,6 +241,13 @@ class Entity extends FluentModel
                     'restaurantReservationsUrl',
                 ],
                 'string'
+            ],
+            [
+                [
+                    'genericImageWidth',
+                    'genericImageHeight',
+                ],
+                'integer'
             ],
             [
                 [